Split integration tests into 3 separate sets to make logs data comparable.
